adds models folder, moves tour schema and model to tourModel.js
