+++ controllers/tourControllers.js
@@ -1,23 +1,11 @@
-const fs = require('fs');
+const Tour = require('./../models/tourModel');
 
 // convert data to JSON object
-const tours = JSON.parse(
-    fs.readFileSync(`${__dirname}/../dev-data/data/tours-simple.json`)
-);
+// const tours = JSON.parse(
+//     fs.readFileSync(`${__dirname}/../dev-data/data/tours-simple.json`)
+// );
 
 /*********** TOUR ROUTE CONTROLLERS **********/
-
-exports.checkID = (req, res, next, val) => {
-    console.log(`Tour ID is ${val}.`);
-    if (req.params.id * 1 > tours.length) {
-        // without return it will continue to next middleware
-        return res.status(404).json({
-            status: 'Fail',
-            message: 'Invalid ID',
-        });
-    };
-    next();
-};
 
 exports.checkBody = (req, res, next) => {
     const body = req.body;
@@ -35,32 +23,32 @@ exports.getAllTours = (req, res) => {
     console.log('Get All Tours Route Hit.');
     res.status(200).json({
         status: 'Success',
-        requestTime: req.requestTime,
-        data: {
-            tours: tours,
-        }
+        // requestTime: req.requestTime,
+        // data: {
+        //     tours: tours,
+        // }
     });
 };
 
 exports.postTour = (req, res) => {
     console.log(req.body);
     // get last tour, get its id, add 1 to create new tour id
-    const newId = tours[tours.length - 1].id + 1;
+    // const newId = tours[tours.length - 1].id + 1;
     // combine id with new tour object
-    const newTour = Object.assign({ id: newId }, req.body);
+    // const newTour = Object.assign({ id: newId }, req.body);
 
-    tours.push(newTour);
+    // tours.push(newTour);
 
-    fs.writeFile(`${__dirname}/../dev-data/data/tours-simple.json`,
-        JSON.stringify(tours),
-        err => {
-            res.status(201).json({
-                status: 'Success',
-                data: {
-                    tour: newTour
-                }
-            });
-        });
+    // fs.writeFile(`${__dirname}/../dev-data/data/tours-simple.json`,
+    //     JSON.stringify(tours),
+    //     err => {
+    //         res.status(201).json({
+    //             status: 'Success',
+    //             data: {
+    //                 tour: newTour
+    //             }
+    //         });
+    //     });
 };
 
 exports.getTour = (req, res) => {
@@ -69,15 +57,15 @@ exports.getTour = (req, res) => {
     const id = req.params.id * 1;
 
     // check if matching id in tours
-    const tour = tours.find(elem => elem.id === id);
-    if (tour) {
-        res.status(200).json({
-            status: 'Success',
-            data: {
-                tour: tour,
-            }
-        });
-    };
+    // const tour = tours.find(elem => elem.id === id);
+    // if (tour) {
+    //     res.status(200).json({
+    //         status: 'Success',
+    //         data: {
+    //             tour: tour,
+    //         }
+    //     });
+    // };
 };
 
 exports.patchTour = (req, res) => {
