+++ models/tourModel.js
@@ -1,0 +1,21 @@
+const mongoose = require('mongoose');
+
+const tourSchema = new mongoose.Schema({
+    name: {
+        type: String,
+        required: [true, 'A tour must have a name'], // validator, checks if name is there
+        unique: true
+    },
+    rating: {
+        type: Number,
+        default: 4.5
+    },
+    price: {
+        type: Number,
+        required: true
+    }
+});
+
+const Tour = mongoose.model('Tour', tourSchema);
+
+module.exports = Tour;
